+++ bin/replicante/src/components/core_api/apply/appliers.rs
@@ -9,12 +9,14 @@ use replicante_stream_events::Stream;
 
 use super::agent_action;
 use super::discovery_settings;
+use super::namespace;
 use super::orchestrator_action;
 use crate::Result;
 
 const APIV_REPLI_V0: &str = "replicante.io/v0";
 const KIND_AGENT_ACTION: &str = "AgentAction";
 const KIND_DISCOVERY_SETTING: &str = "DiscoverySettings";
+const KIND_NAMESPACE: &str = "Namespace";
 const KIND_NODE_ACTION: &str = "NodeAction";
 const KIND_ORCHESTRATOR_ACTION: &str = "OrchestratorAction";
 
@@ -40,6 +42,9 @@ pub fn find(object: &ApplyObject) -> Option<Applier> {
         }
         (APIV_REPLI_V0, KIND_DISCOVERY_SETTING) => {
             Some(Box::new(discovery_settings::replicante_io_v0))
+        }
+        (APIV_REPLI_V0, KIND_NAMESPACE) => {
+            Some(Box::new(namespace::replicante_io_v0))
         }
         (APIV_REPLI_V0, KIND_ORCHESTRATOR_ACTION) => {
             Some(Box::new(orchestrator_action::replicante_io_v0))