+++ bin/replicante/src/components/core_api/apply/mod.rs
@@ -28,6 +28,7 @@ mod agent_action;
 mod appliers;
 mod discovery_settings;
 mod metrics;
+mod namespace;
 mod orchestrator_action;
 mod validate;