+++ bin/replicante/src/components/core_api/mod.rs
@@ -10,6 +10,7 @@ mod apply;
 mod catalogue;
 mod cluster;
 mod discovery_settings;
+mod namespace;
 
 pub use apply::register_metrics;
 
@@ -22,10 +23,12 @@ impl CoreAPI {
         let catalogue = self::catalogue::configure(&logger, interfaces);
         let cluster = self::cluster::configure(&logger, interfaces);
         let discovery_settings = self::discovery_settings::configure(&logger, interfaces);
+        let namespace = self::namespace::configure(&logger, interfaces);
         interfaces.api.configure(apply);
         interfaces.api.configure(catalogue);
         interfaces.api.configure(cluster);
         interfaces.api.configure(discovery_settings);
+        interfaces.api.configure(namespace);
         CoreAPI {}
     }
 }