+++ store/primary/src/backend/mongo/persist.rs
@@ -21,6 +21,7 @@ use replicante_models_core::agent::Node as NodeModel;
 use replicante_models_core::agent::Shard as ShardModel;
 use replicante_models_core::cluster::discovery::DiscoverySettings as DiscoverySettingsModel;
 use replicante_models_core::cluster::ClusterSettings as ClusterSettingsModel;
+use replicante_models_core::scope::Namespace as NamespaceModel;
 
 use super::super::PersistInterface;
 use super::constants::COLLECTION_ACTIONS;
@@ -30,6 +31,7 @@ use super::constants::COLLECTION_AGENTS_INFO;
 use super::constants::COLLECTION_CLUSTER_SETTINGS;
 use super::constants::COLLECTION_DISCOVERIES;
 use super::constants::COLLECTION_DISCOVERY_SETTINGS;
+use super::constants::COLLECTION_NAMESPACES;
 use super::constants::COLLECTION_NODES;
 use super::constants::COLLECTION_SHARDS;
 use super::document::ActionDocument;
@@ -175,6 +177,21 @@ impl PersistInterface for Persist {
         let document = match document {
             Bson::Document(document) => document,
             _ => panic!("DiscoverySettings failed to encode as BSON document"),
+        };
+        replace_one(collection, filter, document, span, self.tracer.as_deref())
+            .with_context(|_| ErrorKind::MongoDBOperation)?;
+        Ok(())
+    }
+
+    fn namespace(&self, namespace: NamespaceModel, span: Option<SpanContext>) -> Result<()> {
+        let filter = doc! {
+            "ns_id": &namespace.ns_id,
+        };
+        let collection = self.client.database(&self.db).collection(COLLECTION_NAMESPACES);
+        let document = bson::to_bson(&namespace).with_context(|_| ErrorKind::MongoDBBsonEncode)?;
+        let document = match document {
+            Bson::Document(document) => document,
+            _ => panic!("Namespace failed to encode as BSON document"),
         };
         replace_one(collection, filter, document, span, self.tracer.as_deref())
             .with_context(|_| ErrorKind::MongoDBOperation)?;