+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Store `Namespace` objects in the primary store.
+
 ### Changed
 - **BREAKING**: Configuration `sentry_capture_api` is now an on/off boolean.
 - **BREAKING**: Migrate `ClusterDiscovery` to `replisdk` provided model.